+++ src/mocks/comments-mock.js
@@ -1,0 +1,10 @@
+export default [{
+  id: 1,
+  user: {
+    id: Math.random(),
+    name: `Kate Muir`
+  },
+  rating: 8.9,
+  comment: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
+  date: `2019-05-08T14:13:56.569Z`
+}];
